requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
redis
python-dotenv
pydantic
requests
google-generativeai
groq
google-auth-oauthlib
google-api-python-client
bcrypt
PyJWT
rapidfuzz
